% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_engine.R
\name{print.bayesian_meta_transport}
\alias{print.bayesian_meta_transport}
\title{Print a Bayesian meta-transport fit}
\usage{
\method{print}{bayesian_meta_transport}(x, ...)
}
\arguments{
\item{x}{A fitted \code{"bayesian_meta_transport"} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a Bayesian meta-transport fit
}
