% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{read_example_targets}
\alias{read_example_targets}
\title{Read the bundled example targets dataset}
\usage{
read_example_targets()
}
\value{
A data frame of target modifier profiles.
}
\description{
Read the bundled example targets dataset
}
